requirements.txt
fastapi
uvicorn
httpx
python-dotenv
gunicorn
aiogram
apscheduler
alembic
redis
pandas
